Refactor MapStoreInsn class
